+++ starkbank/corporaterule/__init__.py
@@ -1,0 +1,2 @@
+from .__corporaterule import CorporateRule
+from .__corporaterule import parse_rules